pyproject.toml
[project]
name = "zarr-testcard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "numcodecs>=0.13.1",
    "pillow>=11.0.0",
    "sqlalchemy>=2.0.36",
]
